Allow to override refresh event publisher
Consumer of SCC Monitor can now override bean RefreshEventPublisher to
implement custom logic of refreshing the connected clients.

The reason for including empty spring.binders is specific of Spring
Cloud Stream binder disovery. If a particular client wants to disable
Cloud Bus at all and override RefreshEventPublisher, application will
not start without any any binder in the classpath (even if it is not
userd). This is a bit hacky solution, probably it would be better to
split Cloud Bus into separate project and do not bring Spring Cloud
Stream as a dependency in Monitor.